requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
banks==2.4.1
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
dotenv==0.9.9
filelock==3.25.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2024.12.0
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
# Core
llama-index-core
llama-index-llms-groq
# Database
neo4j

# Utilities
python-dotenv
langsmith==0.7.13
llama-index-core==0.14.15
llama-index-graph-stores-neo4j==0.6.0
llama-index-instrumentation==0.4.3
llama-index-llms-groq==0.4.1
llama-index-llms-openai==0.6.26
llama-index-llms-openai-like==0.5.3
llama-index-workflows==2.15.0
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
neo4j==5.28.3
neo4j-graphrag==1.13.1
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.2
openai==2.26.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.7.5
pytest==9.0.2
pytest-mock==3.15.1
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scipy==1.17.1
setuptools==82.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.0
tokenizers==0.22.2
tqdm==4.67.3
transformers==4.57.6
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
